pyproject.toml
[project]
name = "modemm"
version = "0.0.1"
authors = [
  { name="Ednaordinary", email="imnotednamode+git@gmail.com" },
]
description = "Explorative node based generative interface"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "diffusers>=0.33.1",
    "fastapi[standard]>=0.115.12",
    "pdoc>=15.0.3",
    "pillow>=11.2.1",
    "transformers>=4.51.3",
]
license = "MIT"
license-files = ["LICEN[CS]E*"]

[project.urls]
Homepage = "https://github.com/Ednaordinary/ModeMM"
Issues = "https://github.com/Ednaordinary/ModeMM/issues"

[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

# Default ruff config
[tool.ruff]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

line-length = 88
indent-width = 4

target-version = "py39"

[tool.ruff.lint]

select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]

quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = false

docstring-code-line-length = "dynamic"
